+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Released]
+
+## [0.3.0] - 2022-02-20
+
+### Added
+
+- Start-CountDown
+- Stop-CountDown
+- Reset-CountDown
+- Set-ScoreBoard
+- Set-Noise
+
 ## [0.2.1] Released
 
 Version bump for Release